Check the two lazy bitmaps, and stop the malloc wrapper truncating
Two small holes in utility.c, both in the one place where they are least
affordable.

gbitmap_create_blank's result went straight into gbitmap_get_data and
then a write, with nothing between. On a failed allocation that is a null
dereference, and these are the aplite and diorite shading bitmaps: built
lazily during the first render rather than at init, which is exactly when
the heap has least room, on the platform which has 24k of it for code as
well. Every other allocation here goes through MALLOC, which halts
loudly; these did not check at all. They now draw nothing and try again
on the next pass, because losing the grey is better than losing the app.

malloc_check took a uint16_t. Nothing in this app asks for anywhere near
64k, so it was never reached, but it is the wrong type on the one
function whose entire job is to fail loudly rather than quietly: any
request of 65536 or more would have been truncated to size % 65536 and
handed back as a buffer smaller than the caller asked for, which is a
heap overflow written by the safety net. It takes a size_t now, and the
four host test stubs of it move with it.

+++ src/c/utility.h
@@ -54,7 +54,7 @@ void graphics_fill_rect_grey_light(GContext *ctx, GRect rect);
 //! @param size The size of the memory to allocate
 //! @param file The name of the file it is called from
 //! @param line The line number it is called from
-void *malloc_check(uint16_t size, const char *file, int line);
+void *malloc_check(size_t size, const char *file, int line);
 
 //! Get current epoch in milliseconds
 //! @return The current epoch time in milliseconds
